Refactor: Folder Hierarquie, Rollback job, refresh setup docs

+++ scripts/__init__.py
@@ -4,8 +4,10 @@ from loguru import logger
 from pathlib import Path
 
 pkg_path = [x for x in sys.path if x.endswith('site-packages')]
-with open(fr'{pkg_path[0]}/my_paths.pth', 'w') as file:
-    file.write(str(Path.cwd()))
+
+if not Path(fr'{pkg_path[0]}/my_paths.pth').exists():
+    with open(fr'{pkg_path[0]}/my_paths.pth', 'w') as file:
+        file.write(str(Path.cwd()))
 
 
 config = {
@@ -13,9 +15,11 @@ config = {
         {
             "sink": sys.stderr,
             "format": (
-                "\n<level>===========================================</level>\n"
-                "<level>{message}</level>\n"
-                "<level>===========================================</level>\n"
+                "<level>\n"
+                "===========================================\n"
+                "{message}\n"
+                "===========================================\n"
+                "</level>"
             ),
         }
     ]
